Add files via upload
I have created a UI design and prototypes using Figma for an app called "The Hungry Hunter," which enables users to order a variety of junk food items at any time, as the service operates 24/7. The app's interface includes a front screen, login and sign-up pages, and a crucial home screen. The home screen allows users to mark items as favorites, modify their profile, and access the menu. Various additional options are available as well. Users can select items from the menu, add them to their cart, and proceed to the cart and checkout screens. These screens offer a range of features. Within the checkout screen, users can even adjust the delivery location. The final step involves a confirmation screen. This summarizes the core features and flow of the project.